+++ projects/aca-shared/src/lib/components/page-layout/page-layout.component.html
@@ -1,0 +1,3 @@
+<ng-content select="aca-page-layout-header"></ng-content>
+<ng-content select="aca-page-layout-error" *ngIf="hasError"></ng-content>
+<ng-content select="aca-page-layout-content" *ngIf="!hasError"></ng-content>
