+++ projects/aca-shared/tslint.json
@@ -1,17 +1,7 @@
 {
   "extends": "../../tslint.json",
   "rules": {
-    "directive-selector": [
-      true,
-      "attribute",
-      "lib",
-      "camelCase"
-    ],
-    "component-selector": [
-      true,
-      "element",
-      "lib",
-      "kebab-case"
-    ]
+    "directive-selector": [true, "attribute", "aca", "camelCase"],
+    "component-selector": [true, "element", "aca", "kebab-case"]
   }
 }
