+++ src/app/components/about/about.component.html
@@ -1,5 +1,5 @@
-<app-page-layout>
-  <app-page-layout-content [scrollable]="true">
+<aca-page-layout>
+  <aca-page-layout-content [scrollable]="true">
     <div class="main-content">
       <article>
         <header>{{ 'application.name' | adfAppConfig }}</header>
@@ -43,5 +43,5 @@
         <app-package-list [data]="dependencyEntries"></app-package-list>
       </article>
     </div>
-  </app-page-layout-content>
-</app-page-layout>
+  </aca-page-layout-content>
+</aca-page-layout>
