+++ src/app/components/favorite-libraries/favorite-libraries.component.html
@@ -1,5 +1,5 @@
-<app-page-layout>
-  <app-page-layout-header>
+<aca-page-layout>
+  <aca-page-layout-header>
     <adf-breadcrumb root="APP.BROWSE.LIBRARIES.MENU.FAVORITE_LIBRARIES.TITLE">
     </adf-breadcrumb>
 
@@ -8,9 +8,9 @@
         <aca-toolbar-action [actionRef]="entry"></aca-toolbar-action>
       </ng-container>
     </adf-toolbar>
-  </app-page-layout-header>
+  </aca-page-layout-header>
 
-  <app-page-layout-content>
+  <aca-page-layout-content>
     <div class="main-content">
       <adf-document-list
         #documentList
@@ -91,5 +91,5 @@
     <div class="sidebar" *ngIf="(infoDrawerOpened$ | async)">
       <aca-info-drawer [node]="selection.last"></aca-info-drawer>
     </div>
-  </app-page-layout-content>
-</app-page-layout>
+  </aca-page-layout-content>
+</aca-page-layout>
