--- src/app/components/layout/page-layout/page-layout.component.html
@@ -1,3 +1,0 @@
-<ng-content select="app-page-layout-header"></ng-content>
-<ng-content select="app-page-layout-error" *ngIf="hasError"></ng-content>
-<ng-content select="app-page-layout-content" *ngIf="!hasError"></ng-content>
