--- src/app/components/layout/page-layout/page-layout.theme.scss
@@ -1,46 +1,0 @@
-@mixin app-page-layout-theme($theme) {
-  $foreground: map-get($theme, foreground);
-
-  .app-page-layout {
-    @include flex-column;
-
-    .app-page-layout-header {
-      display: flex;
-      align-items: center;
-      flex: 0 0 65px;
-      flex-basis: 48px;
-      background: #fafafa;
-      border-bottom: 1px solid mat-color($foreground, text, 0.07);
-      padding: 0 24px;
-    }
-
-    .app-page-layout-content {
-      @include flex-row;
-    }
-
-    .app-page-layout-error {
-      @include flex-row;
-    }
-
-    .main-content {
-      @include flex-column;
-      border-right: 1px solid mat-color($foreground, text, 0.07);
-    }
-
-    .scrollable {
-      overflow: auto !important;
-
-      .main-content {
-        overflow: auto !important;
-      }
-    }
-
-    .sidebar {
-      display: block;
-      height: 100%;
-      overflow-y: scroll;
-      max-width: 350px;
-      width: 350px;
-    }
-  }
-}
