+++ src/app/components/search/search-libraries-results/search-libraries-results.component.html
@@ -1,14 +1,14 @@
-<app-page-layout>
-  <app-page-layout-header>
+<aca-page-layout>
+  <aca-page-layout-header>
     <adf-breadcrumb root="APP.BROWSE.SEARCH_LIBRARIES.TITLE"> </adf-breadcrumb>
     <adf-toolbar class="adf-toolbar--inline">
       <ng-container *ngFor="let entry of actions; trackBy: trackByActionId">
         <aca-toolbar-action [actionRef]="entry"></aca-toolbar-action>
       </ng-container>
     </adf-toolbar>
-  </app-page-layout-header>
+  </aca-page-layout-header>
 
-  <app-page-layout-content>
+  <aca-page-layout-content>
     <div class="main-content">
       <div class="adf-search-results">
         <div class="adf-search-results__content">
@@ -123,5 +123,5 @@
     <div class="sidebar" *ngIf="(infoDrawerOpened$ | async)">
       <aca-info-drawer [node]="selection.last"></aca-info-drawer>
     </div>
-  </app-page-layout-content>
-</app-page-layout>
+  </aca-page-layout-content>
+</aca-page-layout>
